Move SafeContext to separate file
