+++ src/hooks/useAuthenticate.test.ts
@@ -4,7 +4,7 @@ import { SafeClient } from '@safe-global/sdk-starter-kit'
 import { useAuthenticate } from '@/hooks/useAuthenticate.js'
 import { renderHookInMockedSafeProvider } from '@test/utils.js'
 import { signerPrivateKeys } from '@test/fixtures.js'
-import { SafeContextType } from '@/SafeProvider.js'
+import { SafeContextType } from '@/SafeContext.js'
 
 describe('useAuthenticate', () => {
   const signerClientMock = { safeClient: 'signer' } as unknown as SafeClient
