+++ src/hooks/usePublicClient.ts
@@ -1,7 +1,7 @@
 import { useContext, useEffect, useState } from 'react'
 import { type SafeClient } from '@safe-global/sdk-starter-kit'
 import type { ConfigParam, SafeConfig } from '@/types/index.js'
-import { SafeContext } from '@/SafeProvider.js'
+import { SafeContext } from '@/SafeContext.js'
 import { useCompareObject } from '@/hooks/helpers/useCompare.js'
 import { createPublicClient } from '@/createClient.js'
