+++ src/index.ts
@@ -1,4 +1,5 @@
 export { useConfirmTransaction, useSafe, useSendTransaction } from '@/hooks/index.js'
 export * from '@/types/index.js'
-export { SafeProvider, SafeContext } from './SafeProvider.js'
+export { SafeProvider } from './SafeProvider.js'
+export { SafeContext } from './SafeContext.js'
 export { createConfig } from './createConfig.js'
